master <- dev

+++ ui/src/views/networks/list/index.js
@@ -1,20 +1,21 @@
-import React, { useEffect, useState } from 'react'
-import styled from 'styled-components'
+import { useMutation, useQuery } from '@apollo/client'
 import { navigate, useLocation } from '@reach/router'
-import { useQuery, useMutation } from 'react-apollo'
-
-import { GET_NETWORKS, DELETE_NETWORK } from '_graphql/actions/networks'
-
-import { Dragon as Spinner } from '_components/spinner'
-import SearchInput from '_components/inputs/search-input'
-import ErrorState from '_components/error-state'
-import EmptyState from '_components/empty-state'
-import Button from '_components/button'
-import toast from '_components/toast'
-import Text from '_components/text'
-import Flex from '_components/flex'
+import React, { useEffect, useState } from 'react'
+import { FixedSizeList } from 'react-window'
+import InfiniteLoader from 'react-window-infinite-loader'
+import styled from 'styled-components'
 import Box from '_components/box'
-
+import Button from '_components/button'
+import { useConfirmationDialog } from '_components/confirmation-dialog'
+import EmptyState from '_components/empty-state'
+import ErrorState from '_components/error-state'
+import Flex from '_components/flex'
+import SearchInput from '_components/inputs/search-input'
+import { Dragon as Spinner } from '_components/spinner'
+import Text from '_components/text'
+import { DELETE_NETWORK } from '_graphql/mutations'
+import { GET_NETWORKS } from '_graphql/queries'
+import { useToast } from '_utils/toast-provider'
 import NetworkCard from './network-card'
 
 const Container = styled(Flex)`
@@ -23,19 +24,18 @@ const Container = styled(Flex)`
 
 const NetworksView = () => {
   const [searchFilter, setSearchFilter] = useState('')
-
   const location = useLocation()
+
+  const { success } = useToast()
+  const { confirm } = useConfirmationDialog()
 
   const getNetworksQuery = useQuery(GET_NETWORKS, {})
 
   const [deleteNetwork, deleteNetworkMutation] = useMutation(DELETE_NETWORK, {
     variables: { id: undefined },
     onCompleted: () => {
-      toast.success('Network deleted')
+      success('Network deleted')
       getNetworksQuery.refetch()
-    },
-    onError: () => {
-      toast.error('Error deleting host')
     },
   })
 
@@ -44,30 +44,42 @@ const NetworksView = () => {
     getNetworksQuery.refetch()
   }, [location])
 
-  const handleNetworkSelect = id => {
+  const handleNetworkSelect = (id) => {
     navigate(`networks/${id}`)
   }
 
-  const handleNetworkDelete = id => {
-    deleteNetwork({ variables: { id } })
+  const handleNetworkDelete = (id) => {
+    confirm({
+      title: 'Are you sure?',
+      details: `This will destroy all resources, including interfaces and connections, associated with this network.`,
+      isDestructive: true,
+      onConfirm: () => {
+        deleteNetwork({ variables: { id } })
+          .then(() => {
+            getNetworksQuery.refetch()
+          })
+          .catch(() => {})
+      },
+    })
   }
 
   const handleCreateNetworkClick = () => {
     navigate('networks/new')
   }
 
+  const networks = getNetworksQuery.data ? getNetworksQuery.data.result : []
+
   const isError = getNetworksQuery.error
   const isLoading = getNetworksQuery.loading || deleteNetworkMutation.loading
-  const isEmpty =
-    getNetworksQuery.data === undefined || getNetworksQuery.data.result === null
-      ? true
-      : getNetworksQuery.data.result.items.length === 0
 
-  const filteredNetworks = isEmpty
-    ? []
-    : getNetworksQuery.data.result.items.filter(
-        el => el.name.includes(searchFilter) || el.ipAddressRange.includes(searchFilter)
-      )
+  const isEmpty = networks.length === 0
+
+  const filteredNetworks = networks.filter(
+    (el) =>
+      el.ID.includes(searchFilter) ||
+      el.Name.includes(searchFilter) ||
+      el.AddressRange.includes(searchFilter)
+  )
 
   if (isLoading) {
     return <Spinner />
@@ -85,7 +97,7 @@ const NetworksView = () => {
         <SearchInput
           width="100%"
           placeholder="Search..."
-          onChange={e => setSearchFilter(e.target.value)}
+          onChange={(value) => setSearchFilter(value)}
           mr={2}
         />
       </Box>
@@ -94,17 +106,41 @@ const NetworksView = () => {
       ) : isEmpty ? (
         <EmptyState message="Oops! It seems that you don't have any networks yet registered." />
       ) : (
-        filteredNetworks.map(n => (
-          <NetworkCard
-            key={n.id}
-            id={n.id}
-            name={n.name}
-            ipAddressRange={n.ipAddressRange}
-            numHosts={n.hosts.count}
-            onClick={handleNetworkSelect}
-            onDelete={() => handleNetworkDelete(n.id)}
-          />
-        ))
+        <InfiniteLoader
+          itemCount={networks.length}
+          isItemLoaded={(index) => index + 1 <= networks.length}
+          loadMoreItems={() => {}}
+        >
+          {({ onItemsRendered, ref }) => (
+            <FixedSizeList
+              className="virtualized-list"
+              height={74 * filteredNetworks.length}
+              itemCount={filteredNetworks.length}
+              itemSize={74}
+              onItemsRendered={onItemsRendered}
+              itemData={filteredNetworks}
+              ref={ref}
+              width={'100%'}
+            >
+              {({ index, style }) => {
+                const network = filteredNetworks[index]
+                return (
+                  <NetworkCard
+                    key={network.ID}
+                    id={network.ID}
+                    name={network.Name}
+                    addressRange={network.AddressRange}
+                    interfacesCount={network.InterfacesCount}
+                    connectionsCount={network.ConnectionsCount}
+                    onClick={handleNetworkSelect}
+                    onDelete={() => handleNetworkDelete(network.ID)}
+                    style={style}
+                  />
+                )
+              }}
+            </FixedSizeList>
+          )}
+        </InfiniteLoader>
       )}
     </Container>
   )
